+++ src/Farmer/Builders/Builders.ServiceBus.fs
@@ -8,26 +8,58 @@ open Namespaces
 open Topics
 open System
 
+module CommonMembers=
+    type IForwardTo<'TState> =
+        abstract member SetForwardTo : 'TState -> ResourceName option -> 'TState
+open CommonMembers
+
 type ServiceBusQueueConfig =
     { Name : ResourceName
+      Namespace : LinkedResource
       LockDuration : TimeSpan option
       DuplicateDetection : TimeSpan option
       DefaultMessageTimeToLive : TimeSpan option
+      ForwardTo : ResourceName option
       Session : bool option
       DeadLetteringOnMessageExpiration : bool option
       MaxDeliveryCount : int option
       MaxSizeInMegabytes : int<Mb> option
       EnablePartitioning : bool option
       AuthorizationRules : Map<ResourceName, AuthorizationRuleRight Set>}
-
+    interface IBuilder with
+      member this.ResourceId = queues.resourceId (this.Namespace.Name/this.Name)
+      member this.BuildResources location = 
+        [
+          { Name = this.Name
+            Namespace = this.Namespace
+            LockDuration = this.LockDuration |> Option.map IsoDateTime.OfTimeSpan
+            DuplicateDetectionHistoryTimeWindow = this.DuplicateDetection |> Option.map IsoDateTime.OfTimeSpan
+            Session = this.Session
+            DeadLetteringOnMessageExpiration = this.DeadLetteringOnMessageExpiration
+            DefaultMessageTimeToLive = this.DefaultMessageTimeToLive |> Option.map IsoDateTime.OfTimeSpan
+            ForwardTo = this.ForwardTo
+            MaxDeliveryCount = this.MaxDeliveryCount
+            MaxSizeInMegabytes = this.MaxSizeInMegabytes
+            EnablePartitioning = this.EnablePartitioning }
+          for rule in this.AuthorizationRules do
+            { QueueAuthorizationRule.Name = rule.Key.Map(fun name -> $"{this.Namespace.Name.Value}/{this.Name.Value}/%s{name}")
+              Location = location
+              Dependencies = [
+                namespaces.resourceId this.Name
+                queues.resourceId (this.Name, this.Name)
+              ]
+              Rights = rule.Value }
+        ]
 type ServiceBusQueueBuilder() =
     member _.Yield _ =
         { Name = ResourceName.Empty
+          Namespace = Managed (namespaces.resourceId ResourceName.Empty)
           LockDuration = None
           DuplicateDetection = None
           Session = None
           DeadLetteringOnMessageExpiration = None
           DefaultMessageTimeToLive = None
+          ForwardTo = None
           MaxDeliveryCount = None
           MaxSizeInMegabytes = None
           EnablePartitioning = None
@@ -60,10 +92,17 @@ type ServiceBusQueueBuilder() =
     /// Add authorization rule on the queue.
     [<CustomOperation "add_authorization_rule">]
     member _.AddAuthorizationRule(state:ServiceBusQueueConfig, name, rights) = { state with AuthorizationRules = state.AuthorizationRules.Add(ResourceName name, Set rights) }
-
+    /// Instead of creating or modifying a namespace, configure this subscription to point to another unmanaged namespace instance.
+    [<CustomOperation "link_to_unmanaged_namespace">]
+    member this.LinkToUnmanagedNamespace (state:ServiceBusQueueConfig, namespaceName:ResourceName) =
+        { state with Namespace = Unmanaged(namespaces.resourceId namespaceName) }
+    member this.LinkToUnmanagedNamespace (state:ServiceBusQueueConfig, namespaceName) =
+        { state with Namespace = Unmanaged(namespaces.resourceId(ResourceName namespaceName)) }
+    interface IForwardTo<ServiceBusQueueConfig> with
+        member this.SetForwardTo state resource = {state with ForwardTo = resource}
 type ServiceBusSubscriptionConfig =
     { Name : ResourceName
-
+      Topic : LinkedResource
       LockDuration : TimeSpan option
       DuplicateDetection : TimeSpan option
       DefaultMessageTimeToLive : TimeSpan option
@@ -71,11 +110,29 @@ type ServiceBusSubscriptionConfig =
       MaxDeliveryCount : int option
       Session : bool option
       DeadLetteringOnMessageExpiration : bool option
-      Rules : Rule list }
+      Rules : Rule list 
+      DependsOn: Set<ResourceId> }
+    interface IBuilder with
+      member this.ResourceId = subscriptions.resourceId (this.Topic.Name/this.Topic.ResourceId.Segments.[0]/this.Name)
+      member this.BuildResources location = 
+        [
+          { Name = this.Name
+            Topic = this.Topic
+            LockDuration = this.LockDuration |> Option.map IsoDateTime.OfTimeSpan
+            DuplicateDetectionHistoryTimeWindow = this.DuplicateDetection |> Option.map IsoDateTime.OfTimeSpan
+            DefaultMessageTimeToLive = this.DefaultMessageTimeToLive |> Option.map IsoDateTime.OfTimeSpan
+            ForwardTo = this.ForwardTo
+            MaxDeliveryCount = this.MaxDeliveryCount
+            Session = this.Session
+            DeadLetteringOnMessageExpiration = this.DeadLetteringOnMessageExpiration
+            Rules = this.Rules 
+            DependsOn = this.DependsOn}
+          ]
 
 type ServiceBusSubscriptionBuilder() =
     member _.Yield _ =
         { Name = ResourceName.Empty
+          Topic = Managed (namespaces.resourceId ResourceName.Empty)
           LockDuration = None
           DuplicateDetection = None
           DefaultMessageTimeToLive = None
@@ -83,7 +140,8 @@ type ServiceBusSubscriptionBuilder() =
           MaxDeliveryCount = None
           Session = None
           DeadLetteringOnMessageExpiration = None
-          Rules = List.empty }
+          Rules = List.empty
+          DependsOn = Set.empty }
 
     /// The name of the queue.
     [<CustomOperation "name">]
@@ -111,8 +169,6 @@ type ServiceBusSubscriptionBuilder() =
     [<CustomOperation "enable_dead_letter_on_message_expiration">]
     member _.DeadLetteringOnMessageExpiration(state:ServiceBusSubscriptionConfig) = { state with DeadLetteringOnMessageExpiration = Some true }
     /// Automatically forward to a queue or topic
-    [<CustomOperation "forward_to">]
-    member _.ForwardTo(state:ServiceBusSubscriptionConfig, target) = { state with ForwardTo = Some (ResourceName target) }
     /// Adds filtering rules for a subscription
     [<CustomOperation "add_filters">]
     member _.AddFilters(state:ServiceBusSubscriptionConfig, filters) = { state with Rules = state.Rules @ filters }
@@ -122,7 +178,16 @@ type ServiceBusSubscriptionBuilder() =
     /// Adds a correlation filtering rule for a subscription
     [<CustomOperation "add_correlation_filter">]
     member this.AddCorrelationFilter(state:ServiceBusSubscriptionConfig, name, properties) = this.AddFilters(state, [ Rule.CreateCorrelationFilter(name, properties) ])
-
+    /// Instead of creating or modifying a namespace, configure this subscription to point to another unmanaged namespace instance.
+    [<CustomOperation "link_to_unmanaged_topic">]
+    member this.LinkToUnmanagedTopic (state:ServiceBusSubscriptionConfig, topicName:ResourceName) =
+        { state with Topic = Unmanaged(topics.resourceId topicName) }
+    member this.LinkToUnmanagedTopic (state:ServiceBusSubscriptionConfig, topicName) =
+        { state with Topic = Unmanaged(topics.resourceId(ResourceName topicName)) }
+    interface IForwardTo<ServiceBusSubscriptionConfig> with
+        member this.SetForwardTo state resource = {state with ForwardTo = resource}
+    interface IDependable<ServiceBusSubscriptionConfig> with
+      member _.Add state resIds = {state with DependsOn = state.DependsOn + resIds }
 
 type ServiceBusTopicConfig =
     { Name : ResourceName
@@ -132,8 +197,9 @@ type ServiceBusTopicConfig =
       EnablePartitioning : bool option
       MaxSizeInMegabytes : int<Mb> option
       Subscriptions : Map<ResourceName, ServiceBusSubscriptionConfig> }
+    member this.ResourceId = topics.resourceId (this.Namespace.Name, this.Name)
     interface IBuilder with
-        member this.ResourceId = topics.resourceId (this.Namespace.Name, this.Name)
+        member this.ResourceId = this.ResourceId
         member this.BuildResources location = [
             { Name = this.Name
               Dependencies = [
@@ -150,21 +216,8 @@ type ServiceBusTopicConfig =
               EnablePartitioning = this.EnablePartitioning
               MaxSizeInMegabytes = this.MaxSizeInMegabytes }
             for subscription in this.Subscriptions do
-                let subscription = subscription.Value
-                { Name = subscription.Name
-                  Namespace =
-                        match this.Namespace with
-                        | Managed resId
-                        | Unmanaged resId -> resId.Name
-                  Topic = this.Name
-                  LockDuration = subscription.LockDuration |> Option.map IsoDateTime.OfTimeSpan
-                  DuplicateDetectionHistoryTimeWindow = subscription.DuplicateDetection |> Option.map IsoDateTime.OfTimeSpan
-                  DefaultMessageTimeToLive = subscription.DefaultMessageTimeToLive |> Option.map IsoDateTime.OfTimeSpan
-                  ForwardTo = subscription.ForwardTo
-                  MaxDeliveryCount = subscription.MaxDeliveryCount
-                  Session = subscription.Session
-                  DeadLetteringOnMessageExpiration = subscription.DeadLetteringOnMessageExpiration
-                  Rules = subscription.Rules }
+                let subscription = {subscription.Value with Topic = Managed this.ResourceId }:> IBuilder
+                yield! subscription.BuildResources location
         ]
 
 type ServiceBusTopicBuilder() =
@@ -206,8 +259,7 @@ type ServiceBusTopicBuilder() =
         { state with Namespace = Unmanaged(namespaces.resourceId namespaceName) }
     member this.LinkToUnmanagedNamespace (state:ServiceBusTopicConfig, namespaceName) =
         { state with Namespace = Unmanaged(namespaces.resourceId(ResourceName namespaceName)) }
-
-
+    
 type ServiceBusConfig =
     { Name : ResourceName
       Sku : Sku
@@ -231,36 +283,12 @@ type ServiceBusConfig =
               Tags = this.Tags  }
 
             for queue in this.Queues do
-              let queue = queue.Value
-              { Name = queue.Name
-                Namespace = this.Name
-                LockDuration = queue.LockDuration |> Option.map IsoDateTime.OfTimeSpan
-                DuplicateDetectionHistoryTimeWindow = queue.DuplicateDetection |> Option.map IsoDateTime.OfTimeSpan
-                Session = queue.Session
-                DeadLetteringOnMessageExpiration = queue.DeadLetteringOnMessageExpiration
-                DefaultMessageTimeToLive =
-                    match queue.DefaultMessageTimeToLive, this.Sku with
-                    | None, Basic -> TimeSpan.FromDays 14.
-                    | None, (Standard | Premium _) -> TimeSpan.MaxValue
-                    | Some ttl, _ -> ttl
-                    |> IsoDateTime.OfTimeSpan
-                MaxDeliveryCount = queue.MaxDeliveryCount
-                MaxSizeInMegabytes = queue.MaxSizeInMegabytes
-                EnablePartitioning = queue.EnablePartitioning }
-              for rule in queue.AuthorizationRules do
-                { QueueAuthorizationRule.Name = rule.Key.Map(fun rule -> $"{this.Name.Value}/{queue.Name.Value}/%s{rule}")
-                  Location = location
-                  Dependencies = [
-                    namespaces.resourceId this.Name
-                    queues.resourceId (this.Name, queue.Name)
-                  ]
-                  Rights = rule.Value }
-
+              let queue = {queue.Value with Namespace = Managed(namespaces.resourceId this.Name)} :> IBuilder
+              yield! queue.BuildResources location
 
             for topic in this.Topics do
                 let topic = {topic.Value with Namespace = Managed(namespaces.resourceId this.Name)} :> IBuilder
-                for topicResource in topic.BuildResources location do
-                    topicResource
+                yield! topic.BuildResources location
 
             for rule in this.AuthorizationRules do
               { Name = rule.Key.Map(fun rule -> $"{this.Name.Value}/%s{rule}")
@@ -322,3 +350,11 @@ let serviceBus = ServiceBusBuilder()
 let topic = ServiceBusTopicBuilder()
 let queue = ServiceBusQueueBuilder()
 let subscription = ServiceBusSubscriptionBuilder()
+
+type IForwardTo<'TState> with
+    [<CustomOperation "forward_to">]
+    member this.ForwardTo(state:'TState, target) = this.SetForwardTo state target
+    member this.ForwardTo(state:'TState, target) = this.SetForwardTo state (Some target)
+    member this.ForwardTo(state:'TState, target) = this.SetForwardTo state (Some (ResourceName target))
+    member this.ForwardTo(state:'TState, target:ServiceBusQueueConfig) = this.SetForwardTo state (Some target.Name)
+    member this.ForwardTo(state:'TState, target:ServiceBusTopicConfig) = this.SetForwardTo state (Some target.Name)