+++ RELEASE_NOTES.md
@@ -4,6 +4,15 @@ Release Notes
 ## vNext
 * Functions: fixed bug preventing references to AppIsnights or storage accounts in other resource groups
 
+## 1.6.19
+* Application Gateways: support for creating application gateways.
+* Container Service (AKS): support for various addons, including the application gateway ingress controller.
+* ExpressRoute: create authorization keys on newly created circuits.
+* Key Vaults: Add keys to new or existing key vaults.
+* ServiceBus: Allow Service Bus Queues/Topics/Subscriptions to be linked to unmanaged namespaces
+* ServiceBus: Allow adding custom dependencies to Subscriptions
+* WebApp/Functions: Adds 'ftp_state' for controlling FTP access for deployments.
+
 ## 1.6.18
 * Resource Groups: Add support for multiple nested deployments targetting the same resource group
 * Resource Groups: Provide input parameters and key vault references to nested deployments.